# docs/devicemanager_swagger.yaml
basePath: /
definitions:
...
  models.Attribute:
    properties:
      key:
        type: string
      origin:
        type: string
      value:
        type: string
    type: object
...
  models.Device:
    properties:
      attributes:
        items:
          $ref: '#/definitions/models.Attribute'
        type: array
      device_type_id:
        type: string
      id:
        type: string
      local_id:
        type: string
      name:
        type: string
      owner_id:
        type: string
    type: object
...
info:
  contact: {}
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Device-Manager API
  version: "0.1"
paths:
...
  /devices:
    delete:
      description: delete multiple devices
      parameters:
      - description: wait for done message in kafka before responding
        in: query
        name: wait
        type: boolean
      - description: ids to be deleted
        in: body
        name: message
        required: true
        schema:
          items:
            type: string
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: delete multiple devices
      tags:
      - delete
      - devices
    get:
      description: list devices
      parameters:
      - description: default 100, will be ignored if 'ids' is set
        in: query
        name: limit
        type: integer
      - description: default 0, will be ignored if 'ids' is set
        in: query
        name: offset
        type: integer
      - description: filter
        in: query
        name: search
        type: string
      - description: default name.asc
        in: query
        name: sort
        type: string
      - description: filter; ignores limit/offset; comma-seperated list
        in: query
        name: ids
        type: string
      - description: filter; comma-seperated list
        in: query
        name: device-type-ids
        type: string
      - description: filter; comma-seperated list; lists elements only if they have
          an attribute key that is in the given list
        in: query
        name: attr-keys
        type: string
      - description: filter; comma-seperated list; lists elements only if they have
          an attribute value that is in the given list
        in: query
        name: attr-values
        type: string
      - description: filter; valid values are 'online', 'offline' and an empty string
          for unknown states
        in: query
        name: connection-state
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Device'
            type: array
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: list devices
      tags:
      - list
      - devices
    post:
      description: create device
      parameters:
      - description: wait for done message in kafka before responding
        in: query
        name: wait
        type: boolean
      - description: element
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/models.Device'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Device'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: create device
      tags:
      - create
      - devices
  /devices/{id}:
    delete:
      description: delete device
      parameters:
      - description: Device Id
        in: path
        name: id
        required: true
        type: string
      - description: wait for done message in kafka before responding
        in: query
        name: wait
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: delete device
      tags:
      - delete
      - devices
    get:
      description: get device
      parameters:
      - description: Device Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Device'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: get device
      tags:
      - get
      - devices
    put:
      description: set device; admins may create new devices but only without using
        the UpdateOnlySameOriginAttributesKey query parameter
      parameters:
      - description: Device Id
        in: path
        name: id
        required: true
        type: string
      - description: wait for done message in kafka before responding
        in: query
        name: wait
        type: boolean
      - description: comma separated list; ensure that no attribute from another origin
          is overwritten
        in: query
        name: update-only-same-origin-attributes
        type: string
      - description: element
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/models.Device'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Device'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: set device
      tags:
      - set
      - devices
  /devices/{id}/attributes:
    put:
      description: set device display name
      parameters:
      - description: Device Id
        in: path
        name: id
        required: true
        type: string
      - description: wait for done message in kafka before responding
        in: query
        name: wait
        type: boolean
      - description: display name
        in: body
        name: message
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Device'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: set device display name
      tags:
      - set
      - devices
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"